# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Interphase Engineers - Project Control & Margin Tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ProjectProjectType =
  (typeof ProjectProjectType)[keyof typeof ProjectProjectType];

export const ProjectProjectType = {
  electrical: "electrical",
  civil: "civil",
  mechanical: "mechanical",
  plumbing: "plumbing",
  hvac: "hvac",
  other: "other",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Project {
  id: number;
  projectName: string;
  clientName: string;
  projectType: ProjectProjectType;
  location: string;
  projectValue: number;
  startDate: string;
  endDate: string;
  area?: number;
  teamEngineer?: string;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectRequestProjectType =
  (typeof CreateProjectRequestProjectType)[keyof typeof CreateProjectRequestProjectType];

export const CreateProjectRequestProjectType = {
  electrical: "electrical",
  civil: "civil",
  mechanical: "mechanical",
  plumbing: "plumbing",
  hvac: "hvac",
  other: "other",
} as const;

export type CreateProjectRequestStatus =
  (typeof CreateProjectRequestStatus)[keyof typeof CreateProjectRequestStatus];

export const CreateProjectRequestStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface CreateProjectRequest {
  /** @minLength 1 */
  projectName: string;
  /** @minLength 1 */
  clientName: string;
  projectType: CreateProjectRequestProjectType;
  location: string;
  /** @minimum 0 */
  projectValue: number;
  startDate: string;
  endDate: string;
  area?: number;
  teamEngineer?: string;
  status: CreateProjectRequestStatus;
}

export interface ProjectSummary {
  projectId: number;
  projectName: string;
  projectValue: number;
  estimatedCost: number;
  actualCost: number;
  grossMargin: number;
  marginPercent: number;
  materialVariance: number;
  labourVariance: number;
  burnRate: number;
  forecastAtCompletion: number;
}

export interface EstimateRow {
  id: number;
  projectId: number;
  itemCategory: string;
  itemName: string;
  specification?: string;
  unit: string;
  estimatedQuantity: number;
  rate: number;
  cost: number;
  remarks?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEstimateRowRequest {
  projectId: number;
  /** @minLength 1 */
  itemCategory: string;
  /** @minLength 1 */
  itemName: string;
  specification?: string;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  estimatedQuantity: number;
  /** @minimum 0 */
  rate: number;
  remarks?: string;
}

export interface ConsumptionEntry {
  id: number;
  projectId: number;
  weekNumber: number;
  weekStart: string;
  weekEnd: string;
  itemName: string;
  estimatedQty: number;
  actualQty: number;
  variance: number;
  variancePercent: number;
  unitCost: number;
  actualCost: number;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConsumptionEntryRequest {
  projectId: number;
  weekNumber: number;
  weekStart: string;
  weekEnd: string;
  /** @minLength 1 */
  itemName: string;
  /** @minimum 0 */
  estimatedQty: number;
  /** @minimum 0 */
  actualQty: number;
  /** @minimum 0 */
  unitCost: number;
  notes?: string;
}

export interface CategoryBreakdownItem {
  category: string;
  estimatedCost: number;
  actualCost: number;
  variance: number;
  variancePercent: number;
}

export interface WeeklyTrendPoint {
  weekNumber: number;
  weekLabel: string;
  estimatedCost: number;
  actualCost: number;
  cumulativeEstimated: number;
  cumulativeActual: number;
}

export interface MarginReport {
  projectId: number;
  projectName: string;
  projectValue: number;
  estimatedCost: number;
  actualCost: number;
  grossMargin: number;
  marginPercent: number;
  materialVariance: number;
  labourVariance: number;
  burnRate: number;
  forecastAtCompletion: number;
  categoryBreakdown: CategoryBreakdownItem[];
  weeklyTrend: WeeklyTrendPoint[];
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  totalRevenue: number;
  totalEstimatedCost: number;
  totalActualCost: number;
  overallMarginPercent: number;
  projectsAtRisk: number;
  recentProjects: Project[];
}

export type ListProjectsParams = {
  status?: string;
  search?: string;
};

export type ListEstimatesParams = {
  projectId: number;
  category?: string;
};

export type ListConsumptionParams = {
  projectId: number;
  weekNumber?: number;
};

export type GetMarginReportParams = {
  projectId: number;
};

export type GetWeeklyTrendParams = {
  projectId: number;
};

export type GetCategoryBreakdownParams = {
  projectId: number;
};