# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Interphase Engineers - Project Control & Margin Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consumption:
    get:
      operationId: listConsumption
      tags: [consumption]
      summary: List weekly consumption entries
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
        - name: weekNumber
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Consumption entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionEntry"
    post:
      operationId: createConsumptionEntry
      tags: [consumption]
      summary: Create consumption entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsumptionEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /consumption/{id}:
    put:
      operationId: updateConsumptionEntry
      tags: [consumption]
      summary: Update consumption entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsumptionEntryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionEntry"
    delete:
      operationId: deleteConsumptionEntry
      tags: [consumption]
      summary: Delete consumption entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    ConsumptionEntry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        weekNumber:
          type: integer
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        itemName:
          type: string
        estimatedQty:
          type: number
          format: double
        actualQty:
          type: number
          format: double
        variance:
          type: number
          format: double
        variancePercent:
          type: number
          format: double
        unitCost:
          type: number
          format: double
        actualCost:
          type: number
          format: double
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - weekNumber
        - weekStart
        - weekEnd
        - itemName
        - estimatedQty
        - actualQty
        - variance
        - variancePercent
        - unitCost
        - actualCost
        - createdAt
        - updatedAt
    CreateConsumptionEntryRequest:
      type: object
      properties:
        projectId:
          type: integer
        weekNumber:
          type: integer
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        itemName:
          type: string
          minLength: 1
        estimatedQty:
          type: number
          format: double
          minimum: 0
        actualQty:
          type: number
          format: double
          minimum: 0
        unitCost:
          type: number
          format: double
          minimum: 0
        notes:
          type: string
      required:
        - projectId
        - weekNumber
        - weekStart
        - weekEnd
        - itemName
        - estimatedQty
        - actualQty
        - unitCost
...