# lib/api-zod/src/generated/types/consumptionEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Interphase Engineers - Project Control & Margin Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumptionEntry {
  id: number;
  projectId: number;
  weekNumber: number;
  weekStart: Date;
  weekEnd: Date;
  itemName: string;
  estimatedQty: number;
  actualQty: number;
  variance: number;
  variancePercent: number;
  unitCost: number;
  actualCost: number;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}