# lib/api-zod/src/generated/types/createConsumptionEntryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Interphase Engineers - Project Control & Margin Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateConsumptionEntryRequest {
  projectId: number;
  weekNumber: number;
  weekStart: Date;
  weekEnd: Date;
  /** @minLength 1 */
  itemName: string;
  /** @minimum 0 */
  estimatedQty: number;
  /** @minimum 0 */
  actualQty: number;
  /** @minimum 0 */
  unitCost: number;
  notes?: string;
}